% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_sdg_keywords}
\alias{count_sdg_keywords}
\title{Count SDG keywords in a given text}
\usage{
count_sdg_keywords(
  text,
  goal_num,
  keywords = "elsevier100",
  count_repeats = FALSE
)
}
\arguments{
\item{text}{The target text to check SDG keywords}

\item{goal_num}{The SDG goal we intend to check keywords for}

\item{keywords}{The specific data set from which to draw keywords (ex. "elsevier", "sdsn")}

\item{count_repeats}{Whether or not to count repeats for the keywords}
}
\value{
A number representing the weight of words for an SDG
}
\description{
Determine the total weights of all SDG keywords that exist in a given text
for a given goal.
}
\examples{
count_goal("Poverty entails more than the lack of income and productive resources
to ensure sustainable livelihoods.", 1)
}
