% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_sdg_weights}
\alias{count_sdg_weights}
\title{Count SDG keywords}
\usage{
count_sdg_weights(text, sdg, keywords = "elsevier100", count_repeats = FALSE)
}
\arguments{
\item{text}{The target text to map SDG keywords}

\item{sdg}{The SDG goal we intend to check keywords for}

\item{keywords}{The specific data set from which to draw keywords (ex. "elsevier", "sdsn")}

\item{count_repeats}{Whether or not to count repeat weights for the keywords}
}
\value{
A number representing the total weight of words for an SDG
}
\description{
For a given SDG, determine the total weights of the keywords
that exist in the text.
}
\examples{
count_goal(text = "Poverty entails more than the lack of income and productive resources
to ensure sustainable livelihoods.", sdg = 1)
}
