% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_sdgs_weights}
\alias{count_sdgs_weights}
\title{Count SDGs keywords}
\usage{
count_sdgs_weights(text, keywords = "elsevier2", count_repeats = FALSE)
}
\arguments{
\item{text}{The target text to check SDGs keywords}

\item{keywords}{The specific data set from which to draw keywords (ex. "elsevier", "sdsn")}

\item{count_repeats}{Whether or not to count repeats for the keywords}
}
\value{
A number representing the weight of words for all SDGs
}
\description{
Determine the total weights of all SDGs keywords that exist in a given text
for a given goal.
}
\examples{
count_sdgs_keywords("Poverty entails more than the lack of income and productive resources
to ensure sustainable livelihoods.")
}
