% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_sdgs_keywords}
\alias{map_sdgs_keywords}
\title{Map text to SDGs keywords}
\usage{
map_sdgs_keywords(df, keywords = "elsevier100", count_repeats = FALSE)
}
\arguments{
\item{df}{Dataframe with a column `text`}

\item{keywords}{The specific data set from which to draw keywords (ex. "elsevier", "sdsn")}

\item{count_repeats}{Whether or not to count repeats for the keywords}
}
\value{
Original Dataframe plus 16 additional columns representing vectors of
keywords for SDGs 1 to 16.
}
\description{
Return vector of keywords for SDGs 1 to 16
}
\examples{
map_sdgs_keywords(df, keywords = "elsevier", count_repeats = TRUE)
}
