% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_sdgs_weights}
\alias{map_sdgs_weights}
\title{Map text to all SDGs}
\usage{
map_sdgs_weights(df, keywords = "elsevier100", count_repeats = FALSE)
}
\arguments{
\item{df}{Dataframe with a column `text`}

\item{keywords}{The specific data set from which to draw keywords (ex. "elsevier", "sdsn")}

\item{count_repeats}{Whether or not to count repeats for the keywords}
}
\value{
Original dataframe plus 16 additional columns representing weights for
SDGs 1 to 16.
}
\description{
Calculate the total weight for all SDGs from 1 to 16
}
\examples{
map_sdgs_weights(df, keywords = "elsevier", count_repeats = TRUE)
}
